requirements.txt
langgraph>=0.2.0
langchain>=0.2.10
langchain-openai>=0.1.23
langchain-ollama>=0.1.1
meraki>=1.56.0
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.32.0
google-search-results>=2.4.2   # SerpAPI
typer>=0.12.3
rich>=13.7.1
pandas>=2.2.2
fastapi>=0.112
uvicorn[standard]>=0.30
apscheduler>=3.10